# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VoiceTranscribeRequest {
  /** Base64-encoded audio data (M4A, WAV, MP3, WebM) */
  audio: string;
  /** BCP-47 language code hint for Whisper (e.g. "es", "en"). Optional — auto-detects if omitted. */
  language?: string;
}

export interface VoiceTranscribeResponse {
  /** Transcribed text from user audio */
  text: string;
}

/**
 * Who sent the message
 */
export type ConversationMessageRole =
  (typeof ConversationMessageRole)[keyof typeof ConversationMessageRole];

export const ConversationMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ConversationMessage {
  /** Who sent the message */
  role: ConversationMessageRole;
  /** Text content of the message */
  text: string;
}

/**
 * gpt-4o-audio-preview voice for assistant response
 */
export type VoiceChatRequestVoice =
  (typeof VoiceChatRequestVoice)[keyof typeof VoiceChatRequestVoice];

export const VoiceChatRequestVoice = {
  alloy: "alloy",
  echo: "echo",
  fable: "fable",
  onyx: "onyx",
  nova: "nova",
  shimmer: "shimmer",
} as const;

export interface VoiceChatRequest {
  /** Base64-encoded audio data (M4A, WAV, MP3, WebM) */
  audio: string;
  /** gpt-4o-audio-preview voice for assistant response */
  voice?: VoiceChatRequestVoice;
  /** BCP-47 language code hint for Whisper transcription (e.g. "es", "en", "fr"). Optional. */
  language?: string;
  /** Recent conversation history to provide context to the AI (up to 20 messages). Optional. */
  history?: ConversationMessage[];
}

export interface VoiceChatResponse {
  /** Base64-encoded MP3 audio response from gpt-4o-audio-preview */
  audio: string;
  /** Whisper transcription of user's audio input */
  userText: string;
  /** Text of the gpt-4o-audio-preview assistant response */
  assistantText: string;
}

export interface VoiceError {
  error: string;
}