# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: voice
    description: AI Voice Assistant operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /voice/transcribe:
    post:
      operationId: voiceTranscribe
      tags: [voice]
      summary: Transcribe audio to text
      description: Uses Whisper (gpt-4o-mini-transcribe) to convert base64-encoded audio into text.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceTranscribeRequest"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceTranscribeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceError"
  /voice/chat:
    post:
      operationId: voiceChat
      tags: [voice]
      summary: Full speech-to-speech AI conversation turn
      description: |
        Runs a full voice conversation turn in parallel:
        1. Whisper (gpt-4o-mini-transcribe) → user transcript
        2. gpt-4o-audio-preview → assistant response text + MP3 audio
        Returns both transcripts and the base64-encoded MP3.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceChatRequest"
      responses:
        "200":
          description: AI voice response with audio and transcripts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    VoiceTranscribeRequest:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data (M4A, WAV, MP3, WebM)
        language:
          type: string
          description: BCP-47 language code hint for Whisper (e.g. "es", "en"). Optional — auto-detects if omitted.
      required:
        - audio
    VoiceTranscribeResponse:
      type: object
      properties:
        text:
          type: string
          description: Transcribed text from user audio
      required:
        - text
    ConversationMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
          description: Who sent the message
        text:
          type: string
          description: Text content of the message
      required:
        - role
        - text
    VoiceChatRequest:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data (M4A, WAV, MP3, WebM)
        voice:
          type: string
          enum: [alloy, echo, fable, onyx, nova, shimmer]
          description: gpt-4o-audio-preview voice for assistant response
        language:
          type: string
          description: BCP-47 language code hint for Whisper transcription (e.g. "es", "en", "fr"). Optional.
        history:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
          description: Recent conversation history to provide context to the AI (up to 20 messages). Optional.
      required:
        - audio
    VoiceChatResponse:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded MP3 audio response from gpt-4o-audio-preview
        userText:
          type: string
          description: Whisper transcription of user's audio input
        assistantText:
          type: string
          description: Text of the gpt-4o-audio-preview assistant response
      required:
        - audio
        - userText
        - assistantText
    VoiceError:
      type: object
      properties:
        error:
          type: string
      required:
        - error