# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Uses Whisper (gpt-4o-mini-transcribe) to convert base64-encoded audio into text.
 * @summary Transcribe audio to text
 */
export const VoiceTranscribeBody = zod.object({
  audio: zod
    .string()
    .describe("Base64-encoded audio data (M4A, WAV, MP3, WebM)"),
  language: zod
    .string()
    .optional()
    .describe(
      'BCP-47 language code hint for Whisper (e.g. \"es\", \"en\"). Optional — auto-detects if omitted.',
    ),
});

export const VoiceTranscribeResponse = zod.object({
  text: zod.string().describe("Transcribed text from user audio"),
});

/**
 * Runs a full voice conversation turn in parallel:
1. Whisper (gpt-4o-mini-transcribe) → user transcript
2. gpt-4o-audio-preview → assistant response text + MP3 audio
Returns both transcripts and the base64-encoded MP3.

 * @summary Full speech-to-speech AI conversation turn
 */
export const VoiceChatBody = zod.object({
  audio: zod
    .string()
    .describe("Base64-encoded audio data (M4A, WAV, MP3, WebM)"),
  voice: zod
    .enum(["alloy", "echo", "fable", "onyx", "nova", "shimmer"])
    .optional()
    .describe("gpt-4o-audio-preview voice for assistant response"),
  language: zod
    .string()
    .optional()
    .describe(
      'BCP-47 language code hint for Whisper transcription (e.g. \"es\", \"en\", \"fr\"). Optional.',
    ),
  history: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]).describe("Who sent the message"),
        text: zod.string().describe("Text content of the message"),
      }),
    )
    .optional()
    .describe(
      "Recent conversation history to provide context to the AI (up to 20 messages). Optional.",
    ),
});

export const VoiceChatResponse = zod.object({
  audio: zod
    .string()
    .describe("Base64-encoded MP3 audio response from gpt-4o-audio-preview"),
  userText: zod
    .string()
    .describe("Whisper transcription of user's audio input"),
  assistantText: zod
    .string()
    .describe("Text of the gpt-4o-audio-preview assistant response"),
});