# lib/api-zod/src/generated/types/conversationMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessageRole } from "./conversationMessageRole";

export interface ConversationMessage {
  /** Who sent the message */
  role: ConversationMessageRole;
  /** Text content of the message */
  text: string;
}