# lib/api-zod/src/generated/types/voiceChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessage } from "./conversationMessage";
import type { VoiceChatRequestVoice } from "./voiceChatRequestVoice";

export interface VoiceChatRequest {
  /** Base64-encoded audio data (M4A, WAV, MP3, WebM) */
  audio: string;
  /** gpt-4o-audio-preview voice for assistant response */
  voice?: VoiceChatRequestVoice;
  /** BCP-47 language code hint for Whisper transcription (e.g. "es", "en", "fr"). Optional. */
  language?: string;
  /** Recent conversation history to provide context to the AI (up to 20 messages). Optional. */
  history?: ConversationMessage[];
}